+++ test/selenium-grid-wct.conf.js
@@ -56,7 +56,7 @@ module.exports = {
         {
           "browserName":  "safari",
           "platform":     "OS X 10.11",
-          "version":      "9"
+          "version":      "10"
         }
       ]
     },