+++ wct.conf.js
@@ -36,7 +36,7 @@ module.exports = {
         {
           "browserName":  "safari",
           "platform":     "OS X 10.11",
-          "version":      "9"
+          "version":      "10"
         }
       ]
     },